requirements.txt
beautifulsoup4
requests
pymongo
fastapi
uvicorn
pydantic[dotenv]
kafka-python
fastapi-utils